add edit & publish routes to document handler
