+++ src/routes/v1/edit.route.ts
@@ -1,58 +1,17 @@
 import { Elysia, t } from 'elysia';
-import { DataValidator } from '../../classes/DataValidator';
 import { ErrorSender } from '../../classes/ErrorSender';
-import { DocumentManager } from '../../classes/DocumentManager';
-import { basePath, maxDocLength } from '../../utils/constants.ts';
+import { DocumentHandler } from '../../classes/DocumentHandler.ts';
 
 export default new Elysia({
 	name: 'routes:v1:documents:edit'
 }).patch(
 	':id',
-	async ({ request, body, params: { id } }) => {
-		if (!DataValidator.isAlphanumeric(id))
-			return ErrorSender.sendError(400, {
-				type: 'error',
-				errorCode: 'jsp.invalid_input',
-				message: 'Invalid ID provided'
-			});
-
-		const file = Bun.file(basePath + id);
-
-		const fileExists = await file.exists();
-
-		if (!fileExists)
-			return ErrorSender.sendError(400, {
-				type: 'error',
-				errorCode: 'jsp.file_not_found',
-				message: 'The requested file does not exist'
-			});
-
-		const buffer = Buffer.from(body as ArrayBuffer);
-
-		if (!DataValidator.isLengthBetweenLimits(buffer, 1, maxDocLength)) {
-			return ErrorSender.sendError(400, {
-				type: 'error',
-				errorCode: 'jsp.invalid_file_length',
-				message: 'The document data its outside of max length or is null'
-			});
-		}
-
-		const doc = await DocumentManager.read(file);
-
-		if (doc.secret != request.headers.get('secret')) {
-			return ErrorSender.sendError(403, {
-				type: 'error',
-				errorCode: 'jsp.invalid_secret',
-				message: 'The secret is not correct'
-			});
-		}
-
-		doc.rawFileData = buffer;
-
-		await DocumentManager.write(basePath + id, doc);
-
-		return { message: 'File updated successfully' };
-	},
+	async ({ request, body, params: { id } }) =>
+		DocumentHandler.handleEdit({
+			id,
+			newBody: body,
+			secret: request.headers.get('secret') || ''
+		}),
 	{
 		type: 'arrayBuffer',
 		body: t.Any({ description: 'The new file' }),
