+++ src/utils/constants.ts
@@ -1,7 +1,7 @@
 import type { ServerOptions } from '../interfaces/ServerOptions.ts';
 
 export const defaultServerOptions: ServerOptions = {
-	hostname: process.env['HOSTNAME'] ?? 'https://jspaste.eu',
+	hostname: process.env['HOSTNAME'] || 'https://jspaste.eu',
 	port: process.env['PORT'] ?? 4000,
 
 	// FIXME: Fix correct order when v2 is finished
@@ -9,5 +9,5 @@ export const defaultServerOptions: ServerOptions = {
 } as const satisfies Required<ServerOptions>;
 
 // TODO: Move to Server as static?
-export const basePath = process.env['DOCUMENTS_PATH'] ?? 'documents/';
-export const maxDocLength = parseInt(process.env['MAX_FILE_LENGTH'] ?? '2000000');
+export const basePath = process.env['DOCUMENTS_PATH'] || 'documents/';
+export const maxDocLength = parseInt(process.env['MAX_FILE_LENGTH'] || '2000000');
